filenames.icn is the merge of changes made to clean up a few of the intial assumptions made and the addition of code to handle compressed ICode.
The file auto.y and the changes made to udb.icn are the addition of
several tools to aid in debugging unicon programs.  These tools are not
well documented yet, but the documentation is coming.  Unless they are
explicitly enabled by the user they will not affect the exectution of
the source level debugger.


git-svn-id: http://svn.code.sf.net/p/unicon/code@1253 e5ec3061-cdee-4f97-8dda-ba8bb3dbdd71