Don't add the port to the HOST field in the HTTP header unless explicitly set by the the user
Some HTTP servers are picky (buggy?) and don't like to see the optional
port appended to the HOST field inside the HTTP header. For this reason,
keep the port out by default. This allows a client to respond to a 301
error by following the returned location field to the 'letter'.
More fine tuning might be required to cover various cases.



git-svn-id: http://svn.code.sf.net/p/unicon/code@4486 e5ec3061-cdee-4f97-8dda-ba8bb3dbdd71